fix(runtime): nao trunca o exit code do filho para int8
tk_rt_run/tk_rt_run_quiet devolviam (int32_t)(int8_t)WEXITSTATUS(...)
(4 sitios, 2 funcoes x POSIX/_WIN32), truncando qualquer exit code
128..255 para negativo — e.g. 232 vira -24. Isso torna 128 dos 256
exit codes POSIX inexpressaveis num cenario `.tkr`: declarar o valor
honesto (232) falha a asserção porque o runner reporta -24; declarar
-24 falha AINDA PIOR porque mf_read_int engole o '-' e devolve 0 sem
erro (M.3 silencioso). scripts/*.sh legados comparam `$?` (unsigned,
sempre corretos); a consolidacao .tkr regrediria essa cobertura sem
este fix.

Remove o cast (int8_t) intermediario nos 4 sitios; o valor faz
round-trip 0..255 corretamente agora, e nada abaixo de 128 muda
(mudanca estritamente monotona — os 12 call-sites de
teko::process::run/run_quiet usam apenas ==0/!=0/pass-through).

mf_read_int (src/build/manifest.tks) permanece intocado: sua
assinatura -> u64 serve os pisos de cobertura TOML, onde um '-' seria
tipo-errado.

+++ src/runtime/teko_rt.c
@@ -1672,7 +1672,7 @@ int32_t tk_rt_run(const tk_str *argv, uint64_t n) {
 #ifdef _WIN32
     // _spawnvp(_P_WAIT) is synchronous: blocks until the child exits, returns its exit code.
     int w = tk_win32_spawnvp(cargv[0], cargv);
-    return (w == -1) ? 127 : (int32_t)(int8_t)w;
+    return (w == -1) ? 127 : (int32_t)w;
 #else
     pid_t pid = fork();
     if (pid < 0) return 127;
@@ -1682,7 +1682,7 @@ int32_t tk_rt_run(const tk_str *argv, uint64_t n) {
     }
     int status = 0;
     if (waitpid(pid, &status, 0) < 0) return 127;
-    if (WIFEXITED(status)) return (int32_t)(int8_t)WEXITSTATUS(status);
+    if (WIFEXITED(status)) return (int32_t)WEXITSTATUS(status);
     return 127;
 #endif
 }
@@ -1712,7 +1712,7 @@ int32_t tk_rt_run_quiet(const tk_str *argv, uint64_t n) {
     _dup2(saved_err, _fileno(stderr));
     _close(saved_out);
     _close(saved_err);
-    return (w == -1) ? 127 : (int32_t)(int8_t)w;
+    return (w == -1) ? 127 : (int32_t)w;
 #else
     pid_t pid = fork();
     if (pid < 0) return 127;
@@ -1728,7 +1728,7 @@ int32_t tk_rt_run_quiet(const tk_str *argv, uint64_t n) {
     }
     int status = 0;
     if (waitpid(pid, &status, 0) < 0) return 127;
-    if (WIFEXITED(status)) return (int32_t)(int8_t)WEXITSTATUS(status);
+    if (WIFEXITED(status)) return (int32_t)WEXITSTATUS(status);
     return 127;
 #endif
 }
